Fix: another quick fix (float conversion)

+++ src/networks/dwa_utils.py
@@ -45,6 +45,6 @@ class Linear_dwa(torch.nn.Module):
     def forward(self, x, m):
         # expand the weights for the relevant process
         mask_weight = torch.mul(self.weight, m).permute(0,2,1)
-        out = torch.einsum("ac,acb->ab", x, mask_weight) + self.bias
+        out = torch.einsum("ac,acb->ab", x.float(), mask_weight.float()) + self.bias
         return out
